Finished subsystem, added IndexANote command
The subsytem's complete, with some help on the constructor from Noah. I also created the IndexANote command, which brings a note that is in contact with the index conveyor up to the sensor's position and then stops. A couple things need to be checked: 1) does the sensor need to be inverted? 2) should I add the intake subsystem to IndexANote, so that it both picks up and indexes a note? Doing so would compact & simplify commands, but also make the intake less independent. Also, a value needs to be defined for the desired speed of the conveyor motor.