Bugfixing, vision index kinda works?, amp unindex also works
So the auton chooser needs to be tested, the vision-based auto note pick up works but if A is released and you drive to a note without enabling AutoIndex() first it'll just pop right back out. Added some bugfixing to shuffleboard, made the LEDs extend SubsystemBase so that they could be added as requirements. Also, the amp indexing now works as a toggle since m_complete is now set to false at the beginning of the command so it can be run over and over.